fb_gcs_to_gbq/requirements.txt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud
google-cloud-bigquery
google-cloud-core
google-cloud-secret-manager==2.0.0
google-cloud-storage
httplib2
oauth2client
oauthlib
pandas
pandas-gbq
requests
requests-oauthlib
fsspec
gcsfs
